requirements.txt
bigquery
pandas
psycopg2
datetime
python-dateutil